Removed roles store
