--- src/store/documents-proposal/getters.js
@@ -1,4 +1,0 @@
-/*
-export function someGetter (state) {
-}
-*/
