--- src/store/documents-proposal/mutations.js
@@ -1,4 +1,0 @@
-/*
-export function someMutation (state) {
-}
-*/
