--- src/store/objects/state.js
@@ -1,5 +1,0 @@
-export default function () {
-  return {
-    //
-  }
-}
