--- src/store/roles/getters.js
@@ -1,2 +1,0 @@
-export const roles = ({ roles }) => roles
-export const proposals = ({ proposals }) => proposals
