--- src/store/roles/state.js
@@ -1,4 +1,0 @@
-export default () => ({
-  roles: [],
-  proposals: []
-})
